Cleaned up code and moved util functions to separate file
